+++ src/components/game/LoadError.tsx
